Add --minify-prefer-performance
Our performance testing has shown that IE suffers a 2-3% degradation
without trailing semicolons in each declaration block -- suggesting IE
treats the semicolon not as a delimiter but as a line-terminator.

This patch is the first step towards allowing developers to choose to
optimize runtime performance over file size depending on their
specific concerns.